fix(editor): change upload and save attachment logic
